26-commit

+++ users/consumers.py
@@ -134,50 +134,52 @@ class GasStationsAsyncWebsocketConsumer(AsyncWebsocketConsumer):
         gas_stations = GasStationModel.objects.all()
         min_distance = float('inf')
         if gas_stations:
+            closest_gas_station = None
             for gas_station in gas_stations:
                 distance = geopy_distance(point, gas_station.point).meters
                 print(self.user.name, distance)
                 if distance <= int(env('DISTANCE')) and distance <= min_distance:
                     min_distance = distance
                     closest_gas_station = gas_station
-            gas_station_user_ = self.user.gas_station_users
-            if gas_station_user_:
-                serializer = UserPointModelSerializer(self.user)
-                gas_station = None
-                gas_station_user = gas_station_user_.filter(gas_station=closest_gas_station).first()
-                if not gas_station_user:
-                    gas_station = gas_station_user.gas_station
-                    gas_station.total-=1
-                    gas_station.save()
+            if closest_gas_station is not None:
+                gas_station_user_ = self.user.gas_station_users
+                if gas_station_user_:
+                    serializer = UserPointModelSerializer(self.user)
+                    gas_station = None
+                    gas_station_user = gas_station_user_.filter(gas_station=closest_gas_station).first()
+                    if not gas_station_user:
+                        gas_station = gas_station_user_.gas_station
+                        gas_station.total-=1
+                        gas_station.save()
 
-                    gas_station = {
-                        'id': str(gas_station.id),
-                        'total': gas_station.total
+                        gas_station = {
+                            'id': str(gas_station.id),
+                            'total': gas_station.total
+                        }
+                        gas_station_user_.delete()
+
+                    message = "Gas Station user updated successfully."
+                    data = {
+                        'action': UPDATE,
+                        'user': serializer.data,
+                        'gas_station': gas_station
                     }
-                    gas_station_user_.delete()
+                else:
+                    GasStationUserModel.objects.create(gas_station=closest_gas_station, user=self.user)
+                    closest_gas_station.total += 1
+                    closest_gas_station.save()
+                    serializer = UserDetailsModelSerializer(self.user)
+                    serializer_gt = GasStationModelSerializer(closest_gas_station)
+                    message = "Gas Station user created successfully."
 
-                message = "Gas Station user updated successfully."
-                data = {
-                    'action': UPDATE,
-                    'user': serializer.data,
-                    'gas_station': gas_station
-                }
-            else:
-                GasStationUserModel.objects.create(gas_station=closest_gas_station, user=self.user)
-                closest_gas_station.total += 1
-                closest_gas_station.save()
-                serializer = UserDetailsModelSerializer(self.user)
-                serializer_gt = GasStationModelSerializer(closest_gas_station)
-                message = "Gas Station user created successfully."
+                    data = {
+                        "action": CREATE,
+                        "user": serializer.data,
+                        "gas_station": serializer_gt.data
+                    }
 
-                data = {
-                    "action": CREATE,
-                    "user": serializer.data,
-                    "gas_station": serializer_gt.data
-                }
-
-            return message, data
-        return None, None
+                return message, data
+            return None, None
 
     @database_sync_to_async
     def create_comment(self, data):
